+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 #### 功能构建
 
-- (无)
+- 改进数据访问层。
+  - com.dwarfeng.subgrade.impl.dao.MemorySingleObjectDao。
 
 #### Bug修复